Move ShaderPreview classes to baaahs.gl.preview.
